Renaming IMovementPolicy and it is now the general AI interface
