+++ Runtime/RPG/AI/Movement/IMovementBehaviour.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0bc6aa2fdcf56dd409b9b0dfc888048a
+guid: f32e39dac6dc50642b4b40ed73abdeb3
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
